+++ CHANGELOG.md
@@ -48,6 +48,8 @@
 - Fixed `ControlPlane` cluster wide resources not migrating to new ownership labels
   (introduced in 1.3.0) when upgrading the operator form 1.2 (or older) to 1.3.0.
   [#369](https://github.com/Kong/gateway-operator/pull/369)
+- Requeue instead of reporting an error when a finalizer removal yields a conflict.
+  [#454](https://github.com/Kong/gateway-operator/pull/454)
 
 ## [v1.3.0]